requirements.txt
# Minimal dependencies for the web UI
Flask>=2.2
flask-socketio>=5.3
python-socketio>=5.8
python-engineio>=4.5
# eventlet is optional; we're using threading async_mode. Uncomment if desired.
# eventlet>=0.33
